refactor: Extract NoticeRepository and add HybridNoticeRepository for notice domain - noticeRepository.ts: Firebase CRUD + NoticePost/NoticeAuthor 타입 정의 분리 - hybridNoticeRepository.ts: server-first 읽기 + Firebase primary 쓰기 (기존 hybrid 패턴 적용) - index.ts: HybridNoticeRepository를 NoticeRepository로 export - noticeService.ts: Firestore/apiClient 직접 의존 제거, 유효성 검사만 유지하고 NoticeRepository 위임
