+++ src/services/noticeService.ts
@@ -1,102 +1,6 @@
-import {
-  addDoc,
-  collection,
-  deleteDoc,
-  doc,
-  getDocs,
-  limit,
-  orderBy,
-  query,
-  serverTimestamp,
-  Timestamp,
-  updateDoc
-} from 'firebase/firestore';
-import { db } from '../config/firebase';
-import { serverRead, serverWrite } from '../data/apiClient';
-import { ServerNoticeRepository } from '../repositories/server/serverNoticeRepository';
+import { NoticeAuthor, NoticePost, NoticeRepository } from '../repositories';
 
-export interface NoticePost {
-  id: string;
-  title: string;
-  content: string;
-  authorName: string;
-  createdAtText: string;
-  commentCount: number;
-}
-
-export interface NoticeAuthor {
-  authorName: string;
-  authorEmail: string;
-}
-
-interface FirestoreNoticeDoc {
-  title?: string;
-  postTitle?: string;
-  content?: string;
-  authorName?: string;
-  authorEmail?: string;
-  createdAt?: unknown;
-  commentCount?: number;
-  comments?: unknown[];
-}
-
-const formatDate = (value: unknown): string => {
-  if (!value) return '-';
-
-  let parsedDate: Date | null = null;
-
-  if (value instanceof Date) {
-    parsedDate = value;
-  } else if (value instanceof Timestamp) {
-    parsedDate = value.toDate();
-  } else if (
-    typeof value === 'object' &&
-    value !== null &&
-    'toDate' in value &&
-    typeof (value as { toDate: () => Date }).toDate === 'function'
-  ) {
-    parsedDate = (value as { toDate: () => Date }).toDate();
-  }
-
-  if (!parsedDate) return '-';
-
-  return parsedDate
-    .toLocaleDateString('ko-KR', {
-      year: 'numeric',
-      month: '2-digit',
-      day: '2-digit'
-    })
-    .replace(/\s/g, '')
-    .replace(/\.$/, '');
-};
-
-const mapServerToNoticePost = (s: import('../repositories/server/serverNoticeRepository').ServerNoticeResponse): NoticePost => ({
-  id: s.id,
-  title: s.title || '(제목 없음)',
-  content: s.content || '',
-  authorName: s.author_name || '-',
-  createdAtText: formatDate(s.created_at ? new Date(s.created_at) : null),
-  commentCount: s.comment_count
-});
-
-const mapDocToNoticePost = (docSnap: { id: string; data: () => FirestoreNoticeDoc }): NoticePost => {
-  const data = docSnap.data();
-  const commentCount =
-    typeof data.commentCount === 'number'
-      ? data.commentCount
-      : Array.isArray(data.comments)
-        ? data.comments.length
-        : 0;
-
-  return {
-    id: docSnap.id,
-    title: data.title || data.postTitle || '(제목 없음)',
-    content: data.content || '',
-    authorName: data.authorName || '-',
-    createdAtText: formatDate(data.createdAt),
-    commentCount
-  };
-};
+export type { NoticePost, NoticeAuthor };
 
 export class NoticeService {
   static async createNoticePost(
@@ -104,98 +8,26 @@ export class NoticeService {
     payload: { title: string; content: string },
     author?: NoticeAuthor
   ): Promise<void> {
-    if (!boxName) {
-      throw new Error('박스 이름이 없습니다.');
-    }
-
+    if (!boxName) throw new Error('박스 이름이 없습니다.');
     const title = payload.title.trim();
     const content = payload.content.trim();
-
-    if (!title) {
-      throw new Error('공지 제목을 입력해주세요.');
-    }
-
+    if (!title) throw new Error('공지 제목을 입력해주세요.');
     const authorName = author?.authorName?.trim();
-    if (!authorName) {
-      throw new Error('작성자를 입력해주세요.');
-    }
-
-    const now = new Date().toISOString();
-    const docRef = await addDoc(collection(db, `box/${boxName}/notices`), {
-      title,
-      content,
+    if (!authorName) throw new Error('작성자를 입력해주세요.');
+    await NoticeRepository.create(boxName, { title, content }, {
       authorName,
-      authorEmail: author?.authorEmail || '',
-      commentCount: 0,
-      comments: [],
-      createdAt: serverTimestamp(),
-      updatedAt: serverTimestamp()
+      authorEmail: author!.authorEmail || '',
     });
-
-    serverWrite(
-      () => ServerNoticeRepository.createNotice({
-        id: docRef.id,
-        box_name: boxName,
-        title,
-        content,
-        author_name: authorName,
-        author_email: author?.authorEmail || null,
-        created_at: now,
-        updated_at: now
-      }),
-      `Notice.create(${docRef.id})`
-    );
   }
 
   static async getNoticePosts(boxName: string): Promise<NoticePost[]> {
     if (!boxName) return [];
-
-    const serverPosts = await serverRead(
-      async () => {
-        const list = await ServerNoticeRepository.listByBox(boxName, 200);
-        return list.map(mapServerToNoticePost);
-      },
-      `Notice.getNoticePosts(${boxName})`
-    );
-    if (serverPosts && serverPosts.length > 0) return serverPosts;
-
-    try {
-      const noticesQuery = query(
-        collection(db, `box/${boxName}/notices`),
-        orderBy('createdAt', 'desc')
-      );
-      const snapshot = await getDocs(noticesQuery);
-      return snapshot.docs.map((docSnap) => mapDocToNoticePost(docSnap));
-    } catch (error) {
-      console.error('Failed to load notice posts:', error);
-      return [];
-    }
+    return NoticeRepository.list(boxName);
   }
 
   static async getRecentNoticePosts(boxName: string, count = 3): Promise<NoticePost[]> {
     if (!boxName) return [];
-
-    const serverPosts = await serverRead(
-      async () => {
-        const list = await ServerNoticeRepository.listByBox(boxName, count);
-        return list.map(mapServerToNoticePost);
-      },
-      `Notice.getRecentNoticePosts(${boxName})`
-    );
-    if (serverPosts && serverPosts.length > 0) return serverPosts;
-
-    try {
-      const noticesQuery = query(
-        collection(db, `box/${boxName}/notices`),
-        orderBy('createdAt', 'desc'),
-        limit(count)
-      );
-      const snapshot = await getDocs(noticesQuery);
-      return snapshot.docs.map((docSnap) => mapDocToNoticePost(docSnap));
-    } catch (error) {
-      console.error('Failed to load recent notice posts:', error);
-      return [];
-    }
+    return NoticeRepository.list(boxName, count);
   }
 
   static async updateNoticePost(
@@ -203,50 +35,17 @@ export class NoticeService {
     noticeId: string,
     payload: { title: string; content: string; authorName: string }
   ): Promise<void> {
-    if (!boxName) {
-      throw new Error('박스 이름이 없습니다.');
-    }
-
+    if (!boxName) throw new Error('박스 이름이 없습니다.');
     const title = payload.title.trim();
     const content = payload.content.trim();
     const authorName = payload.authorName.trim();
-
-    if (!title) {
-      throw new Error('공지 제목을 입력해주세요.');
-    }
-
-    if (!authorName) {
-      throw new Error('작성자를 입력해주세요.');
-    }
-
-    const now = new Date().toISOString();
-    await updateDoc(doc(db, `box/${boxName}/notices`, noticeId), {
-      title,
-      content,
-      authorName,
-      updatedAt: serverTimestamp()
-    });
-
-    serverWrite(
-      () => ServerNoticeRepository.updateNotice(noticeId, {
-        title,
-        content,
-        author_name: authorName,
-        updated_at: now
-      }),
-      `Notice.update(${noticeId})`
-    );
+    if (!title) throw new Error('공지 제목을 입력해주세요.');
+    if (!authorName) throw new Error('작성자를 입력해주세요.');
+    await NoticeRepository.update(boxName, noticeId, { title, content, authorName });
   }
 
   static async deleteNoticePost(boxName: string, noticeId: string): Promise<void> {
-    if (!boxName) {
-      throw new Error('박스 이름이 없습니다.');
-    }
-
-    await deleteDoc(doc(db, `box/${boxName}/notices`, noticeId));
-    serverWrite(
-      () => ServerNoticeRepository.deleteNotice(noticeId),
-      `Notice.delete(${noticeId})`
-    );
+    if (!boxName) throw new Error('박스 이름이 없습니다.');
+    await NoticeRepository.delete(boxName, noticeId);
   }
 }
